BMP390 data compensation done!!
bmp390 data compensation done in integer math operations due to lack of FPU on NAVIG board (F205), pressure + temp data sending over UART DMA nicely as well

+++ Code/CNTRL/Core/Inc/util.h
@@ -29,8 +29,8 @@ typedef struct {
 } __attribute__((packed)) bno055_data;
 
 typedef struct {
-    uint32_t pressure;
-    uint32_t temperature;
+    uint64_t pressure;
+    int64_t temperature;
 } __attribute__((packed)) bmp390_data;
 
 typedef struct {
